requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pydantic
python-dotenv
pytest
redis
aiohttp